Setup client_spec
