+++ spec/support/addressable_shared_examples.rb
@@ -1,0 +1,6 @@
+shared_examples "addressable" do
+
+  it { should have_one :address }
+  its(:address) { should_not be_nil }
+
+end
